+++ src/singleton/message.h
@@ -34,12 +34,12 @@ public:
     Message();
     void init();
 
-    void showMessage(const QString &strChannel, const QString &strData, MessageCategory eMessageCategory, QString strNick = QString::null, int iTime = 0);
+    void showMessage(const QString &strChannel, const QString &strData, MessageCategory eMessageCategory, QString strNick = QString::null, qint64 iTime = 0);
     void showMessageAll(const QString &strData, MessageCategory eMessageCategory);
     void showMessageActive(const QString &strData, MessageCategory eMessageCategory);
 
 private:
-    void saveMessage(const QString &strChannel, const QString &strData, int iTime, QString strNick = QString::null);
+    void saveMessage(const QString &strChannel, const QString &strData, qint64 iTime, QString strNick = QString::null);
     bool isHideJoinPart(const QString &strChannel, MessageCategory eMessageCategory);
     bool isHighlightMessage(const QString &strMessage);
