+++ src/widgets/chat_view.h
@@ -29,7 +29,7 @@ class ChatView : public QWebView
 public:
     ChatView(const QString &_strChatViewChannel);
     void clearMessages();
-    void displayMessage(const QString &strData, MessageCategory eMessageCategory, int iTime, QString strNick = QString::null);
+    void displayMessage(const QString &strData, MessageCategory eMessageCategory, qint64 iTime, QString strNick = QString::null);
     void refreshCSS();
 
 private:
